Audio: dual-core audio implementation.
We need to resort to adjustable timing between samples.
Normally in an emulator that can run zx_exec() cycle much
faster than the actual emulated computer, this is what happens:

    zx_exec() executed in a few milliseconds
    audio buffer feed, playing and waiting for the next buffer.

So audio samples are generated in a very small amount of time and
the buffer is kept full as it plays.

Here we have an emulator that runs at the same speed as the
Spectrum itself, so samples are generated as they are played. If
the emulator is running slower or faster opcodes, the timing
changes and so the playing speed we should provide to match
emulation speed. So we start with a sensible pause between samples
and adjust it dinaymically based on the fact we find the buffer
ready or not.

A better approach would be to meter the whole waiting+playing loop
for N iterations and have a running average, and then adjust the
playing speed accordingly. The system implemented here is not as
good but almost as good and it's much simpler. For now it's enough :)